refactor: move test config to src folder
